+++ src/components/editor/BasicEditor.tsx
@@ -1,37 +1,19 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useRef } from "react";
+import ChangeFont from "./ChangeFont";
+import ChangeFontSize from "./ChangeFontSize";
 
 const BasicEditor = () => {
   const editorRef = useRef<HTMLDivElement>(null);
-  const fontNameSelectorRef = useRef<HTMLSelectElement>(null);
-  const fontSizeSelectorRef = useRef<HTMLSelectElement>(null);
 
   const editorValue = editorRef.current;
-  const fontSizeList: number[] = [10, 12, 16, 18, 24, 32, 48];
 
   const applyStyle = (style: string): void => {
     if (editorValue) {
       document.execCommand(style);
-      focusEditor();
-    }
-  };
 
-  const changeFontName = (name: string) => {
-    if (fontNameSelectorRef) {
-      document.execCommand("fontName", false, name);
-      focusEditor();
-    }
-  };
-
-  const changeFontSize = (size: number) => {
-    if (fontSizeSelectorRef) {
-      document.execCommand("fontSize", false, size.toString());
-      focusEditor();
-    }
-  };
-
-  const focusEditor = () => {
-    if (editorValue) {
-      editorValue.focus({ preventScroll: true });
+      if (editorValue) {
+        editorValue.focus({ preventScroll: true });
+      }
     }
   };
 
@@ -74,29 +56,8 @@ const BasicEditor = () => {
       >
         UL
       </button>
-      <select
-        id="select-font"
-        ref={fontNameSelectorRef}
-        onChange={(e) => changeFontName(e.target.value)}
-        className={`ml-5 mr-2 my-5 p-1 border border-solid border-black rounded`}
-      >
-        <option value="Inherit">Inherit(default)</option>
-        <option value="Arial">Arial</option>
-        <option value="Times New Roman">Times New Roman</option>
-        <option value="Verdana">Verdana</option>
-      </select>
-      <select
-        id="select-font-size"
-        ref={fontSizeSelectorRef}
-        onChange={(e) => changeFontSize(parseInt(e.target.value))}
-        className={`ml-5 mr-2 my-5 p-1 border border-solid border-black rounded`}
-      >
-        {fontSizeList.map((size, index) => (
-          <option key={index} value={size}>
-            {size}
-          </option>
-        ))}
-      </select>
+      <ChangeFont ref={editorRef} button={buttonStyle} />
+      <ChangeFontSize ref={editorRef } />
       <div
         id="editor"
         ref={editorRef}
